Ensure Call::UpdateDeadline returns an error if the deadline has already exceeded.
This error is passed on to the RecvInitialMetadata complete callback at the server when it receives a new request. This ensures that an incoming call is failed immediately without invoking any registered handlers if the call's deadline was already exceeded when the server receives Initial Metadata.

This may potentially fix b/428185413

PiperOrigin-RevId: 811898358